cleancode authen, friendship

+++ src/main/java/com/phithang/mysocialnetwork/service/Impl/FriendshipService.java
@@ -1,5 +1,7 @@
 package com.phithang.mysocialnetwork.service.Impl;
 
+import com.phithang.mysocialnetwork.dto.FriendshipDto;
+import com.phithang.mysocialnetwork.dto.request.FriendshipRequest;
 import com.phithang.mysocialnetwork.entity.FriendshipEntity;
 import com.phithang.mysocialnetwork.entity.UserEntity;
 import com.phithang.mysocialnetwork.exception.AppException;
@@ -11,6 +13,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -26,19 +29,27 @@ public class FriendshipService implements IFriendshipService {
     private NotificationService notificationService;
 
     @Override
-    public FriendshipEntity save(UserEntity sender, UserEntity receiver) {
+    public FriendshipEntity save(FriendshipRequest request) {
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserEntity sender = userService.findUserByEmail(authentication.getName());
+        UserEntity receiver = userService.findById(request.getReceiverId());
         if (sender == null || receiver == null) {
             throw new AppException(ErrorCode.USER_NOT_EXIST);
+        }
+        if (sender.equals(receiver)) {
+            throw new AppException(ErrorCode.INVALID_REQUEST);
         }
         FriendshipEntity existingFriendship = friendshipRepository.findByUser1AndUser2(sender, receiver);
         if (existingFriendship != null) {
             throw new AppException(ErrorCode.FRIENDSHIP_REQUEST_ALREADY_SENT);
         }
+
         FriendshipEntity friendshipEntity = new FriendshipEntity();
         friendshipEntity.setUser1(sender);
         friendshipEntity.setUser2(receiver);
         friendshipEntity.setStatus("PENDING");
         friendshipEntity.setRequestTimestamp(LocalDateTime.now());
+
         try {
             notificationService.createAndSendNotification(
                     receiver,
@@ -51,19 +62,30 @@ public class FriendshipService implements IFriendshipService {
     }
 
     @Override
-    public boolean accept(UserEntity sender, UserEntity receiver) {
+    public boolean acceptFriendRequest(FriendshipRequest friendshipDto) {
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        String receiverEmail = authentication.getName();
+        UserEntity receiver = userService.findUserByEmail(receiverEmail);
+        UserEntity sender = userService.findById(friendshipDto.getReceiverId());
+        if (sender == null) {
+            return false;
+        }
+
         FriendshipEntity friendshipEntity = friendshipRepository.findByUser1AndUser2(sender, receiver);
         if (friendshipEntity == null) {
-            throw new AppException(ErrorCode.FRIENDSHIP_REQUEST_NOT_FOUND);
+            return false;
         }
+
         friendshipEntity.setStatus("ACCEPTED");
         friendshipEntity.setRequestTimestamp(LocalDateTime.now());
+
         try {
             notificationService.createAndSendNotification(
                     sender,
                     receiver.getFirstname() + " " + receiver.getLastname() + " đã chấp nhận lời mời kết bạn.", null
             );
-            return friendshipRepository.save(friendshipEntity) != null;
+            friendshipRepository.save(friendshipEntity);
+            return true;
         } catch (Exception e) {
             throw new AppException(ErrorCode.FRIENDSHIP_REQUEST_FAILED);
         }
@@ -76,10 +98,20 @@ public class FriendshipService implements IFriendshipService {
     }
 
     @Override
-    public Boolean cancelRequest(FriendshipEntity friendshipEntity) {
-        if (friendshipEntity == null) {
-            throw new AppException(ErrorCode.FRIENDSHIP_REQUEST_NOT_FOUND);
+    public boolean cancelRequest(FriendshipRequest friendshipDto) {
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        String receiverEmail = authentication.getName();
+        UserEntity receiver = userService.findUserByEmail(receiverEmail);
+        UserEntity sender = userService.findById(friendshipDto.getReceiverId());
+        if (sender == null) {
+            return false;
         }
+
+        FriendshipEntity friendshipEntity = friendshipRepository.findByUser1AndUser2(sender, receiver);
+        if (friendshipEntity == null) {
+            return false;
+        }
+
         try {
             friendshipRepository.delete(friendshipEntity);
             return true;
@@ -98,7 +130,7 @@ public class FriendshipService implements IFriendshipService {
     }
 
     @Override
-    public List<FriendshipEntity> findALlRequest() {
+    public List<FriendshipDto> findALlRequest() {
         var authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication == null || !authentication.isAuthenticated()) {
             throw new AppException(ErrorCode.UNAUTHENTICATED);
@@ -108,11 +140,16 @@ public class FriendshipService implements IFriendshipService {
         if (userEntity == null) {
             throw new AppException(ErrorCode.USER_NOT_EXIST);
         }
-        return friendshipRepository.findAllRequests(userEntity);
+        List<FriendshipEntity> friendshipEntities = friendshipRepository.findAllRequests(userEntity);
+        List<FriendshipDto> friendshipDtos = new ArrayList<>();
+        for (FriendshipEntity friendshipEntity : friendshipEntities) {
+            friendshipDtos.add(new FriendshipDto(friendshipEntity));
+        }
+        return friendshipDtos;
     }
 
     @Override
-    public List<FriendshipEntity> findAllFriends() {
+    public List<FriendshipDto> findAllFriends() {
         var authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication == null || !authentication.isAuthenticated()) {
             throw new AppException(ErrorCode.UNAUTHENTICATED);
@@ -122,6 +159,14 @@ public class FriendshipService implements IFriendshipService {
         if (userEntity == null) {
             throw new AppException(ErrorCode.USER_NOT_EXIST);
         }
-        return friendshipRepository.findAllFriends(userEntity);
+        List<FriendshipEntity> friendshipEntities = friendshipRepository.findAllFriends(userEntity);
+        List<FriendshipDto> friendshipDtos = new ArrayList<>();
+        for (FriendshipEntity friendshipEntity : friendshipEntities) {
+            if (friendshipEntity.getUser1().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+                friendshipEntity.setUser1(friendshipEntity.getUser2());
+            }
+            friendshipDtos.add(new FriendshipDto(friendshipEntity));
+        }
+        return friendshipDtos;
     }
 }
